Add --build-data flag to mbed test
